+++ src/main/java/co/netier/RestApp/service/UserService.java
@@ -8,8 +8,8 @@ public interface UserService {
 	
 	public void save(User user);
 	public List<User> findAll();
-	public User find(int id);
-	public void update(int id, User user);
-	public void delete(int id);
+	public User find(long id);
+	public void update(long id, User user);
+	public void delete(long id);
 
 }
